% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggaussian.R
\name{loggaussian}
\alias{loggaussian}
\title{Computing log-transformed 1-d Gaussian pdf}
\usage{
loggaussian(x, mu, Sigma2)
}
\arguments{
\item{x}{sample}

\item{mu}{mean of the Gaussian distribution}

\item{Sigma2}{variance the Gaussian distribution}
}
\value{
1 numeric vector pdf of sample length
}
\description{
This function is used to compute log-transformed Gaussian probability density function
}
\examples{
loggaussian(x=0,mu=0,Sigma2=1)


}
\author{
Jiaying Zhao \email{jyzhao@connect.hku.hk}
}
