% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_sem.R
\name{ng_sem}
\alias{ng_sem}
\title{ng_sem}
\usage{
ng_sem(
  dat,
  nk = 3,
  param.miter = 5000,
  param.error = 1e-06,
  n.cores = 24,
  seed_list = NULL
)
}
\arguments{
\item{dat}{single cell RNA-seq expression data matrix}

\item{nk}{number of Gaussian Components (default:3)}

\item{param.miter}{maximum iterations (default:5000)}

\item{param.error}{error bound (default:1e-6)}

\item{n.cores}{number of working CPU cores (default:24)}

\item{params}{Defalut is 500}

\item{g}{gene to estimate}

\item{g.seed}{seed for reproducibility}

\item{k}{Components}
}
\value{
res: a list w.mat = w.mat,
\itemize{
\item{w.mat} adjacency matrix for GRN
\item{likelihood} a list of length for gene numbers containing likelihood recordings
\item{iteration} a vector of length for gene numbers containing iteration recordings
\item{Comp_Weight} a gene-by-component matrix containing component weights
\item{Comp_Sigma2} a gene-by-component matrix containing component variance
\item{time} CPU time in seconds
}
}
\description{
a wrapper function to perform NG-SEM (CPU parallelization)
}
\author{
Jiaying Zhao, \email{jyzhao@connect.hku.hk}
}
