% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_sem_estimation.R
\name{ng_sem_estimation}
\alias{ng_sem_estimation}
\title{ng_sem_estimation}
\usage{
ng_sem_estimation(X, g, params, g.seed)
}
\arguments{
\item{X}{scRNA-seq expression input}

\item{g}{gene to estimate}

\item{params}{a list of parameters n.K=n.K,n.N=n.N,Data.Y=Data.Y
\itemize{
\item{n.K} number of Gaussian components
\item{n.N} number of samples (T-1)
\item{Data.Y} design matrix
}}

\item{g.seed}{seed for reproducibility}
}
\value{
res_for_gene
\itemize{
\item{Lik_list} likelihood recordings
\item{Iter_num} number of iterations
\item{Comp_w} a vector for the weights of the Gaussian Mixture components
\item{Comp_sig2} a vector for the variances of the Gaussian Mixture components
\item{grn_w_g} grn weights for gene g
}
}
\description{
a wrapper function to perform NG-SEM on a specific gene
}
\author{
Jiyaing Zhao, \email{jyzhao@connect.hku.hk}
}
