% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_sem_expectation.R
\name{ng_sem_expectation}
\alias{ng_sem_expectation}
\title{Expectation step}
\usage{
ng_sem_expectation(residual, model)
}
\arguments{
\item{residual}{residual for SEM y-model.a*x, a vecotr of length n.N(sample length)}

\item{model}{Gaussian Mixture model
\itemize{
\item{mu} a vector for the mean of the Gaussian Mixture components
\item{Sigma2} a vector for the variance of the Gaussian Mixture components
\item{w} a vector for the weights of the Gaussian Mixture components
}}
}
\value{
res
\itemize{
\item{Gamma} a matrix
\item{liklih}
}
}
\description{
This function is used for the E step in the optimization
}
\author{
Jiaying Zhao, \email{jyzhao@connect.hku.hk}
}
