% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_sem_maximization.R
\name{ng_sem_maximization}
\alias{ng_sem_maximization}
\title{Maxinization step}
\usage{
ng_sem_maximization(e.step, Data, residual, model)
}
\arguments{
\item{e.step}{results of E-step}

\item{Data}{a list including Y, y
\itemize{
\item{Y} designed matrix
\item{y} dependent vector
}}

\item{residual}{residual for SEM}

\item{model}{\itemize{
\item{mu} a vector for the mean of the Gaussian Mixture components
\item{Sigma2} a vector for the variance of the Gaussian Mixture components
\item{w} a vector for the weights of the Gaussian Mixture components
}}
}
\value{
model
\itemize{
\item{mu} a vector for the mean of the Gaussian Mixture components
\item{Sigma2} a vector for the variance of the Gaussian Mixture components
\item{w} a vector for the weights of the Gaussian Mixture components
\item{a} adjacent weights for GRN
\item{n.K} number of components for the Gaussian Mixture
}
}
\description{
This function is used for the M step in the optimization
}
\author{
Jiaying Zhao, \email{jyzhao@connect.hku.hk}
}
